Fixed Seperation margin issue
