+++ frontend/src/components/homeComponents/Client.js
@@ -1,0 +1,105 @@
+import React from "react";
+
+function Clients() {
+  return (
+    <>
+      <div className="container" data-aos="fade-up">
+        <div className="section-header">
+          <h3>Our CLients</h3>
+          <p>
+            Sed ut perspiciatis unde omnis iste natus error sit voluptatem
+            accusantium doloremque dere santome nida.
+          </p>
+        </div>
+
+        <div
+          className="row g-0 clients-wrap clearfix"
+          data-aos="zoom-in"
+          data-aos-delay="100"
+        >
+          <div className="col-lg-3 col-md-4 col-xs-6">
+            <div className="client-logo">
+              <img
+                src="assets/img/clients/client-1.png"
+                className="img-fluid"
+                alt=""
+              />
+            </div>
+          </div>
+
+          <div className="col-lg-3 col-md-4 col-xs-6">
+            <div className="client-logo">
+              <img
+                src="assets/img/clients/client-2.png"
+                className="img-fluid"
+                alt=""
+              />
+            </div>
+          </div>
+
+          <div className="col-lg-3 col-md-4 col-xs-6">
+            <div className="client-logo">
+              <img
+                src="assets/img/clients/client-3.png"
+                className="img-fluid"
+                alt=""
+              />
+            </div>
+          </div>
+
+          <div className="col-lg-3 col-md-4 col-xs-6">
+            <div className="client-logo">
+              <img
+                src="assets/img/clients/client-4.png"
+                className="img-fluid"
+                alt=""
+              />
+            </div>
+          </div>
+
+          <div className="col-lg-3 col-md-4 col-xs-6">
+            <div className="client-logo">
+              <img
+                src="assets/img/clients/client-5.png"
+                className="img-fluid"
+                alt=""
+              />
+            </div>
+          </div>
+
+          <div className="col-lg-3 col-md-4 col-xs-6">
+            <div className="client-logo">
+              <img
+                src="assets/img/clients/client-6.png"
+                className="img-fluid"
+                alt=""
+              />
+            </div>
+          </div>
+
+          <div className="col-lg-3 col-md-4 col-xs-6">
+            <div className="client-logo">
+              <img
+                src="assets/img/clients/client-7.png"
+                className="img-fluid"
+                alt=""
+              />
+            </div>
+          </div>
+
+          <div className="col-lg-3 col-md-4 col-xs-6">
+            <div className="client-logo">
+              <img
+                src="assets/img/clients/client-8.png"
+                className="img-fluid"
+                alt=""
+              />
+            </div>
+          </div>
+        </div>
+      </div>
+    </>
+  );
+}
+
+export default Clients;
