+++ frontend/src/components/homeComponents/Footer.js
