--- frontend/src/components/screencomponents/homecomponent/Team.js
@@ -1,160 +1,0 @@
-import React from "react";
-
-function Team() {
-  return (
-    <>
-      <section id="team">
-          <div className="container" data-aos="fade-up">
-            <div className="section-header">
-              <h3>Team</h3>
-              <p>
-                Sed ut perspiciatis unde omnis iste natus error sit voluptatem
-                accusantium doloremque
-              </p>
-            </div>
-
-            <div className="row">
-              <div
-                className="col-lg-3 col-md-6"
-                data-aos="zoom-out"
-                data-aos-delay="100"
-              >
-                <div className="member">
-                  <img
-                    src="assets/img/team-1.jpg"
-                    className="img-fluid"
-                    alt=""
-                  />
-                  <div className="member-info">
-                    <div className="member-info-content">
-                      <h4>Walter White</h4>
-                      <span>Chief Executive Officer</span>
-                      <div className="social">
-                        <a href="">
-                          <i className="bi bi-twitter"></i>
-                        </a>
-                        <a href="">
-                          <i className="bi bi-facebook"></i>
-                        </a>
-                        <a href="">
-                          <i className="bi bi-instagram"></i>
-                        </a>
-                        <a href="">
-                          <i className="bi bi-linkedin"></i>
-                        </a>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-
-              <div
-                className="col-lg-3 col-md-6"
-                data-aos="zoom-out"
-                data-aos-delay="200"
-              >
-                <div className="member">
-                  <img
-                    src="assets/img/team-2.jpg"
-                    className="img-fluid"
-                    alt=""
-                  />
-                  <div className="member-info">
-                    <div className="member-info-content">
-                      <h4>Sarah Jhonson</h4>
-                      <span>Product Manager</span>
-                      <div className="social">
-                        <a href="">
-                          <i className="bi bi-twitter"></i>
-                        </a>
-                        <a href="">
-                          <i className="bi bi-facebook"></i>
-                        </a>
-                        <a href="">
-                          <i className="bi bi-instagram"></i>
-                        </a>
-                        <a href="">
-                          <i className="bi bi-linkedin"></i>
-                        </a>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-
-              <div
-                className="col-lg-3 col-md-6"
-                data-aos="zoom-out"
-                data-aos-delay="300"
-              >
-                <div className="member">
-                  <img
-                    src="assets/img/team-3.jpg"
-                    className="img-fluid"
-                    alt=""
-                  />
-                  <div className="member-info">
-                    <div className="member-info-content">
-                      <h4>William Anderson</h4>
-                      <span>CTO</span>
-                      <div className="social">
-                        <a href="">
-                          <i className="bi bi-twitter"></i>
-                        </a>
-                        <a href="">
-                          <i className="bi bi-facebook"></i>
-                        </a>
-                        <a href="">
-                          <i className="bi bi-instagram"></i>
-                        </a>
-                        <a href="">
-                          <i className="bi bi-linkedin"></i>
-                        </a>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-
-              <div
-                className="col-lg-3 col-md-6"
-                data-aos="zoom-out"
-                data-aos-delay="400"
-              >
-                <div className="member">
-                  <img
-                    src="assets/img/team-4.jpg"
-                    className="img-fluid"
-                    alt=""
-                  />
-                  <div className="member-info">
-                    <div className="member-info-content">
-                      <h4>Amanda Jepson</h4>
-                      <span>Accountant</span>
-                      <div className="social">
-                        <a href="">
-                          <i className="bi bi-twitter"></i>
-                        </a>
-                        <a href="">
-                          <i className="bi bi-facebook"></i>
-                        </a>
-                        <a href="">
-                          <i className="bi bi-instagram"></i>
-                        </a>
-                        <a href="">
-                          <i className="bi bi-linkedin"></i>
-                        </a>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </section>
-
-    </>
-  );
-}
-
-export default Team;
